+++ src/obsidian/modal/book-modal.tsx
@@ -1,24 +1,35 @@
 import { App, Modal } from 'obsidian'
 import { Book } from '../../bookshelf/book'
-import { createRoot } from 'react-dom/client'
+import { createRoot, Root } from 'react-dom/client'
 import { StrictMode } from 'react'
 import { BookDetails } from '../../component/book-details/book-details'
 
 export class BookModal extends Modal {
-    constructor(app: App, book: Book) {
+    private root: Root
+
+    constructor(
+        app: App,
+        private book: Book,
+    ) {
         super(app)
 
-        this.setTitle(book.metadata.title)
+        this.root = createRoot(this.containerEl.children[1])
 
-        createRoot(this.contentEl).render(
+        this.update()
+    }
+
+    public update(): void {
+        this.setTitle(this.book.metadata.title)
+
+        this.root.render(
             <StrictMode>
                 <BookDetails
-                    book={book}
+                    book={this.book}
                     openNote={async () => {
                         this.close()
 
-                        if (book.note) {
-                            await app.workspace.openLinkText(book.note.basename, '')
+                        if (this.book.note) {
+                            await this.app.workspace.openLinkText(this.book.note.basename, '')
                         }
                     }}
                 />