+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 For a full diff see [`0.5.0...main`][0.5.0...main].
 
+### Changed
+
+- Update book modal on external file changes
+
 ## [`0.5.0`][0.5.0] - 2025-03-25
 
 For a full diff see [`0.4.0...0.5.0`][0.4.0...0.5.0].